requirements_dev.txt
pip==20.0.2
bumpversion==0.5.3
wheel==0.34.2
watchdog==0.10.2
flake8==3.7.9
tox==3.14.5
coverage==5.0.4
Sphinx==2.4.4
sphinx-rtd-theme==0.4.3
twine==3.1.1

pytest==5.4.1
pytest-runner==5.2
pytest-mock==2.0.0
pytest-djangoapp==0.14.0

django>=1.11
django-modeltranslation>=0.13.1,<1.0'
boto3>=1.9,<=1.13'
polib>=1.1.0,<2'
